Isolate watcher xcom logic in a moduel of its own
